update exception handling

+++ router/post.js
@@ -29,11 +29,12 @@ router.get("/list", async (req, res) => {
         const data = await client.query(sql)
         const row = data.rows
 
-        console.log(row[0])
-
-        result.postList.push(row);
+        if (row.length > 0) {
+            result.postList.push(row)
+        } else {
+            result.message = '게시글이 존재하지 않습니다.'
+        }
         res.send(result)
-
     } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
         result.message = err
         res.send(result)
@@ -59,11 +60,15 @@ router.get("/", async (req, res) => {
         const row1 = data1.rows
         const row2 = data2.rows
 
-        result.post.push(row1)
-        result.commentList.push(row2)
-
+        if (row1.length > 0) {
+            result.post.push(row1)
+            if (row2.length > 0) {
+                result.commentList.push(row2)
+            }
+        } else {
+            result.message = '해당 게시글이 존재하지 않습니다.'
+        }
         res.send(result)
-
     } catch(err) { 
         result.message = err
         res.send(result)
@@ -75,23 +80,27 @@ router.post("/", async (req, res) => {
 
     const postTitleValue = req.body.post_title_value
     const postContentValue = req.body.post_content_value
-    const idvalue = req.bodu.id_value
+    const idValue = req.bodu.id_value
 
-    try {
-        await client.connect()
-        
-        const sql = 'INSERT INTO backend.post (postTitle, postContent, userId) VALUES ($1, $2, $3);' // ? 대신 $로 대체 
-        const values = [postTitleValue, postContentValue, idValue]
-
-        await client.query(sql, values)
-
-        result.success = true
-        result.message = "작성 완료"
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
+    if (postTitleValue == '' || postContentValue == '' || idValue == undefined) { // null값 예외처리
+        result.message = "제목 또는 내용을 입력하세요"
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'INSERT INTO backend.post (postTitle, postContent, userId) VALUES ($1, $2, $3);' // ? 대신 $로 대체 
+            const values = [postTitleValue, postContentValue, idValue]
+    
+            await client.query(sql, values)
+    
+            result.success = true
+            result.message = "작성 완료"
+            res.send(result)
+        } catch(err) { 
+            result.message = err
+            res.send(result)
+        }
     }
 })
 
@@ -101,21 +110,24 @@ router.put("/", async (req, res) => {
     const postNum = req.body.post_num
     const postContentValue = req.body.post_content    
     
-    try {
-        await client.connect()
-        
-        const sql = 'UPDATE backend.post SET postContent=$1; WHERE postNum=$2;' 
-        const values = [postContentValue, postNum]
-
-        await client.query(sql, values)
-
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
-    }
+    if (postContentValue == '' || postNum == undefined) { // null값 예외처리
+        result.message = "작성해주세요"
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'UPDATE backend.post SET postContent=$1; WHERE postNum=$2;' 
+            const values = [postContentValue, postNum]
     
+            await client.query(sql, values)
+    
+            res.send(result)
+        } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
+            result.message = err
+            res.send(result)
+        }
+    }
 })
 
 // 게시글 삭제api
@@ -123,42 +135,23 @@ router.delete("/", async (req, res) => {
     
     const postNum = req.body.post_num
 
-    try {
-        await client.connect()
-        
-        const sql = 'DELETE FROM backend.post WHERE postNum=$1;' 
-        const values = postNum
-
-        await client.query(sql)
-
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
-    }
+    if (postNum == undefined) { // undefined값 예외처리
+        result.message = "게시글이 존재하지 않습니다."
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'DELETE FROM backend.post WHERE postNum=$1;' 
+            const values = postNum
     
-})
-
-// 댓글리스트 가져오기
-router.get("/comment-list", async (req, res) => {    
-
-    result.commentList = []
-
-    try {
-        await client.connect() // await 붙여주는
-        
-        const sql = 'SELECT * FROM backend.comment;' // ? 대신 $로 대체 
-        
-        const data = await client.query(sql)
-        const row = data.rows
-
-        result.commentList.push(row);
-        res.send(result)
-
-    } catch(err) { 
-        result.message = err
-        res.send(result)
+            await client.query(sql)
+    
+            res.send(result)
+        } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
+            result.message = err
+            res.send(result)
+        }
     }
 })
 
@@ -168,19 +161,23 @@ router.post("/comment", async (req, res) => {
     const commentContentValue = req.body.comment_content
     const idValue = req.body.id_value
 
-    try {
-        await client.connect()
-        
-        const sql = 'INSERT INTO backend.comment (commentContent, userId) VALUES ($1, $2);' // ? 대신 $로 대체 
-        const values = [commentContentValue, idValue]
-
-        await client.query(sql, values)
-
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
+    if (commentContentValue == '' || idValue == undefined) { // null값 예외처리
+        result.message = "댓글을 입력하세요"
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'INSERT INTO backend.comment (commentContent, userId) VALUES ($1, $2);' // ? 대신 $로 대체 
+            const values = [commentContentValue, idValue]
+    
+            await client.query(sql, values)
+    
+            res.send(result)
+        } catch(err) { 
+            result.message = err
+            res.send(result)
+        }
     }
 })
 
@@ -189,20 +186,24 @@ router.put("/comment", async (req, res) => {
 
     const commentNum = req.body.comment_num
     const commentContentValue = req.body.comment_content    
+
+    if (commentContentValue == '' || commentNum == undefined) { // null값 예외처리
+        result.message = "작성해주세요"
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'UPDATE backend.comment SET commentContent=$1; WHERE commentNum=$2;' 
+            const values = [commentContentValue, commentNum]
     
-    try {
-        await client.connect()
-        
-        const sql = 'UPDATE backend.comment SET commentContent=$1; WHERE commentNum=$2;' 
-        const values = [commentContentValue, commentNum]
-
-        await client.query(sql, values)
-
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
+            await client.query(sql, values)
+    
+            res.send(result)
+        } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
+            result.message = err
+            res.send(result)
+        }
     }
 })
 
@@ -211,19 +212,23 @@ router.delete("/comment", async (req, res) => {
 
     const commentNum = req.body.comment_num
 
-    try {
-        await client.connect()
-        
-        const sql = 'DELETE FROM backend.comment WHERE commentNum=$1;' 
-        const values = commentNum
-
-        await client.query(sql, values)
-
-        res.send(result)
-
-    } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
-        result.message = err
-        res.send(result)
+    if (commentNum == undefined) { // undefined값 예외처리
+        result.message = "댓글이 존재하지 않습니다."
+        return res.send(result)
+    } else {
+        try {
+            await client.connect()
+            
+            const sql = 'DELETE FROM backend.comment WHERE commentNum=$1;' 
+            const values = commentNum
+    
+            await client.query(sql, values)
+    
+            res.send(result)
+        } catch(err) { // 아 어차피 캐로 다 들어가니까 그냥 쭉 쓰는거네 근데 에러부분 뜨는 방식을 잘 모르겠네
+            result.message = err
+            res.send(result)
+        }
     }
 })
